Separated weapon attack initialising from constructor, moved projectile entity storage to projectileBuffer
Separated weapon attack initialising to reduce the amount of paramaters required for construction, and to organise variable setting as the class is now large enough that this is required.
Moved projectile entity storage to buffer so that projectiles no longer disappear while respawning. Will need to pass a player variable for scoring at a later date.